Переделал генерацию коментариев
