+++ src/models/movies.js
@@ -6,6 +6,8 @@ export default class Movies {
     this._filmsData = [];
     this._activeFilterType = FilterType.AllMOVIES;
     this._filterChangeHandlers = null;
+    // this.timeout = setInterval(() => console.log(this._filmsData, 20000));
+    // this.timeout();
   }
 
   setMovies(filmData) {
